requirements.txt
# SMURF PBX runtime dependencies (no PBX libraries — see README).
fastapi==0.136.0
uvicorn[standard]==0.44.0
websockets==16.0
aiosqlite==0.22.1
httpx==0.28.1
jinja2==3.1.6
pydantic>=2,<3
PyJWT==2.12.1
passlib==1.7.4
bcrypt<4
pyotp==2.9.0
qrcode[pil]==8.2
cryptography>=43,<47
python-multipart==0.0.26
aiosmtplib==5.1.0
orjson==3.11.8
aiortc==1.14.0
av>=13,<16
